requirements.txt
feedparser==6.0.11
kafka-python==2.0.2
trafilatura==1.6.3
readability-lxml==0.8.1
beautifulsoup4==4.12.3
langdetect==1.0.9
python-dotenv==1.0.1
psycopg2-binary==2.9.9
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
